+++ spec/config/node-config.js
@@ -4,7 +4,8 @@ TEST = {
   password: "guest",
   url: "ws://localhost:15674/ws",
   badUrl: "ws://localhost:61625",
-  timeout: 2000
+  timeout: 2000,
+  largeMessageSize: 1023 // in KB, in Node total WebSocket frames needs to be lesser than 1MB
 };
 
 WebSocket = require('websocket').w3cwebsocket;